2 interface added :
1. data from T_45 updater.
2. data from T_45 MODEL.
added thread Sensor Communication thread.

+++ Test/src/main/java/stand/util/ExcelPcmFileWriter.java
@@ -59,8 +59,8 @@ public class ExcelPcmFileWriter implements PcmFileWriter {
 				
 				Long timeTest =  (Long) Time.poll();
 				cellTime.setCellValue(LocalTime.ofSecondOfDay(timeTest.longValue()).toString());
-				Integer dataTest = (Integer) data.poll();
-				cell.setCellValue(dataTest.intValue());
+				Float dataTest = (Float) data.poll();
+				cell.setCellValue(dataTest.floatValue());
 			}
 			sheet.autoSizeColumn(rownum);
 		}
